+++ util/src/main/java/net/automatalib/util/automaton/vpa/OneSEVPAs.java
@@ -34,8 +34,8 @@ import net.automatalib.common.util.IntDisjointSets;
 import net.automatalib.common.util.Pair;
 import net.automatalib.common.util.UnionFindRemSP;
 import net.automatalib.common.util.array.ArrayStorage;
+import net.automatalib.util.automaton.minimizer.OneSEVPAMinimizer;
 import net.automatalib.util.automaton.vpa.SPAConverter.ConversionResult;
-import net.automatalib.util.minimizer.OneSEVPAMinimizer;
 import net.automatalib.util.ts.acceptor.AcceptanceCombiner;
 import net.automatalib.word.Word;
 import org.checkerframework.checker.nullness.qual.NonNull;
